services/inference-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow-cpu==2.15.0
boto3==1.34.0
numpy==1.26.0
pytest==7.4.4
httpx==0.26.0